requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
google-generativeai>=0.7.2
langchain-google-genai>=1.0.0
langdetect==1.0.9
SpeechRecognition==3.10.1
pydub==0.25.1
gunicorn==21.2.0